Develop

+++ stadtmeldeapp/src/main/java/com/stadtmeldeapp/service/CategoryService.java
@@ -5,13 +5,18 @@ import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.stadtmeldeapp.CustomExceptions.NotAllowedException;
 import com.stadtmeldeapp.CustomExceptions.NotFoundException;
+import com.stadtmeldeapp.DTO.MainCategoryWithSubCategoriesDTO;
 import com.stadtmeldeapp.Entity.MaincategoryEntity;
+import com.stadtmeldeapp.Entity.ReportingLocationEntity;
 import com.stadtmeldeapp.Entity.SubcategoryEntity;
 import com.stadtmeldeapp.Repository.MaincategoryRepository;
 import com.stadtmeldeapp.Repository.SubcategoryRepository;
 
 import java.util.List;
+import java.util.ArrayList;
+import jakarta.servlet.http.HttpServletRequest;
 
 @Service
 @Transactional
@@ -23,6 +28,8 @@ public class CategoryService {
     @Autowired
     private SubcategoryRepository subCategoryRepository;
 
+    @Autowired
+    private UserService userService;
     public List<MaincategoryEntity> getAllMainCategories() {
         return mainCategoryRepository.findAll();
     }
@@ -32,11 +39,15 @@ public class CategoryService {
                 .orElseThrow(() -> new NotFoundException("Hauptkategorie nicht gefunden"));
     }
 
-    public MaincategoryEntity saveMainCategory(@NonNull MaincategoryEntity mainCategory) {
+    public MaincategoryEntity saveMainCategory(@NonNull MaincategoryEntity mainCategory, HttpServletRequest request) throws NotFoundException, NotAllowedException {
+        ReportingLocationEntity isAdminForLocation = userService.getUserFromRequest(request).getAdminForLocation();
+        if (! mainCategory.getReportingLocationEntity().equals(isAdminForLocation)) throw new NotAllowedException("Keine Berechtigung");
         return mainCategoryRepository.save(mainCategory);
     }
 
-    public void deleteMainCategory(int id) {
+    public void deleteMainCategory(int id, HttpServletRequest request) throws NotFoundException, NotAllowedException {
+        ReportingLocationEntity isAdminForLocation = userService.getUserFromRequest(request).getAdminForLocation();
+        if (!getMainCategoryById(id).getReportingLocationEntity().equals(isAdminForLocation)) throw new NotAllowedException("Keine Berechtigung");
         mainCategoryRepository.deleteById(id);
     }
 
@@ -45,11 +56,16 @@ public class CategoryService {
                 .orElseThrow(() -> new NotFoundException("Unterkategorie nicht gefunden"));
     }
 
-    public SubcategoryEntity saveSubCategory(@NonNull SubcategoryEntity subCategory) {
+    public SubcategoryEntity saveSubCategory(@NonNull SubcategoryEntity subCategory, HttpServletRequest request) throws NotFoundException, NotAllowedException {
+        ReportingLocationEntity isAdminForLocation = userService.getUserFromRequest(request).getAdminForLocation();
+        if (!subCategory.getMaincategoryEntity().getReportingLocationEntity().equals(isAdminForLocation)) throw new NotAllowedException("Keine Berechtigung");
         return subCategoryRepository.save(subCategory);
     }
 
-    public void deleteSubCategory(int id) {
+    public void deleteSubCategory(int id, HttpServletRequest request) throws NotFoundException, NotAllowedException {
+        ReportingLocationEntity isAdminForLocation = userService.getUserFromRequest(request).getAdminForLocation();
+        ReportingLocationEntity subCategoryLocation = getSubCategoryById(id).getMaincategoryEntity().getReportingLocationEntity();
+        if (!subCategoryLocation.equals(isAdminForLocation)) throw new NotAllowedException("Keine Berechtigung");
         subCategoryRepository.deleteById(id);
     }
 
@@ -59,5 +75,18 @@ public class CategoryService {
 
     public List<MaincategoryEntity> getMaincategoriesByLocationName(String reportingLocationName) {
         return mainCategoryRepository.findByReportingLocationEntity_Name(reportingLocationName);
+    }
+
+    public List<MainCategoryWithSubCategoriesDTO> getMainCategoriesWithSubcategoriesByLocationName(String reportingLocationName) {
+        List<MainCategoryWithSubCategoriesDTO> mainCategoryWithSubCategoriesDTOs = new ArrayList<MainCategoryWithSubCategoriesDTO>();
+        
+        List<MaincategoryEntity> mainCategory = getMaincategoriesByLocationName(reportingLocationName);
+        
+        for (MaincategoryEntity maincategoryEntity : mainCategory) {
+            int mainCategoryId = maincategoryEntity.getId();
+            List<SubcategoryEntity> subcategoryEntities = getSubCategoriesByMainCategoryId(mainCategoryId);
+            mainCategoryWithSubCategoriesDTOs.add(new MainCategoryWithSubCategoriesDTO(maincategoryEntity, subcategoryEntities));
+        }
+        return mainCategoryWithSubCategoriesDTOs;
     }
 }
